requirements.txt
flask==2.3.3
transformers==4.35.0
torch==2.1.0
Pillow==10.0.0
python-dotenv==1.0.0 